inet256srv.Params now contain p2p.Swarms instead of p2p.SecureSwarms
The inet256srv implementation is responsible for creating the layer of security, not the daemon, or caller.
A secure swarm will not be wrapped in another secure layer; it will be passed through.
This means that the in memory swarm (a secure swarm) will not have traffic on it encrypted unnecessarily.